+++ CHANGELOG.md
@@ -16,6 +16,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - Made all `validate*` methods on `JSON::Validator` ultimately call `validate!`
 - Updated addressable dependency to 2.4.0
+- Attached failed `uri` or `pathname` to read errors for more meaning
 
 ## [2.6.2] - 2016-05-13